Move speakers.js -> speakers.ts
